Update city coordinate types from byte to int. #58

+++ src/Game.cs
@@ -680,8 +680,8 @@ namespace CivOne
 
 			City city = new(PlayerNumber(player))
 			{
-				X = (byte)x,
-				Y = (byte)y,
+				X = x,
+				Y = y,
 				NameId = nameId
 			};
 			// Order is important here -
